80% of the way towards a working wishlist feature
almost finished a jquery implementation of the wishlist feature for courses.

It features create & destroy actions in the wishlist controller and some jquery that listens to a checkbox on the courses index for adding courses to a wishlist. it's pretty snazzy.

The bug that is frustrating me is that when you click the checkbox it changes before the jquery event is called. it's kinda weird. will fix it later